// models/triggers-module/triggers/types.ts
export interface TriggerInterface {
  id: string
  type: TriggerEntityTypes

  draft: boolean

  name: string
  comment: string | null
  enabled: boolean

  owner: string | null

  // Channel property specific
  device: string
  channel: string
  property: string
  operator: ConditionOperatorTypes
  operand: string

  // Relations
  relationshipNames: Array<string>

  actions: Array<ActionInterface>
  notifications: Array<NotificationInterface>
  conditions: Array<ConditionInterface>

  // Entity transformers
  isEnabled: boolean
  icon: string
  description: string

  isAutomatic: boolean
  isManual: boolean
  isForChannel: boolean
  isDate: boolean
  isTime: boolean
}

// ENTITY TYPES
// ============

export enum TriggerEntityTypes {
  AUTOMATIC = 'triggers-module/trigger-automatic',
  MANUAL = 'triggers-module/trigger-manual',
  CHANNEL_PROPERTY = 'triggers-module/trigger-channel-property',
}